feat(content): show an ad after the main content of the page
