+++ pages/interface/components/ContentList/index.js
@@ -1,6 +1,15 @@
-import { Box, EmptyState, Link, Pagination, PastTime, TabCoinBalanceTooltip, Text, Tooltip } from '@/TabNewsUI';
-import { CommentIcon, LinkExternalIcon } from '@/TabNewsUI/icons';
-import { getDomain, isExternalLink, isTrustedDomain } from 'pages/interface';
+import {
+  AdBanner,
+  Box,
+  EmptyState,
+  Link,
+  Pagination,
+  PastTime,
+  TabCoinBalanceTooltip,
+  Text,
+  Tooltip,
+} from '@/TabNewsUI';
+import { CommentIcon } from '@/TabNewsUI/icons';
 
 export default function ContentList({ ad, contentList: list, pagination, paginationBasePath, emptyStateProps }) {
   const listNumberStart = pagination.perPage * (pagination.currentPage - 1) + 1;
@@ -19,8 +28,14 @@ export default function ContentList({ ad, contentList: list, pagination, paginat
           }}
           key={`content-list-${listNumberStart}`}
           start={listNumberStart}>
-          <Ad ad={ad} />
+          {ad && (
+            <Box as="li" sx={{ display: 'block', gridColumnStart: 2, '::marker': 'none' }}>
+              <AdBanner ad={ad} />
+            </Box>
+          )}
+
           <RenderItems />
+
           <EndOfRelevant />
         </Box>
       ) : (
@@ -146,55 +161,4 @@ export default function ContentList({ ad, contentList: list, pagination, paginat
   function RenderEmptyMessage(props) {
     return <EmptyState title="Nenhum conteúdo encontrado" {...props} />;
   }
-}
-
-function Ad({ ad }) {
-  if (!ad) {
-    return;
-  }
-
-  const link = ad.source_url || `/${ad.owner_username}/${ad.slug}`;
-  const isAdToExternalLink = isExternalLink(link);
-  const domain = isAdToExternalLink ? `(${getDomain(link)})` : '';
-  const title = ad.title.length > 70 ? ad.title.substring(0, 67).trim().concat('...') : ad.title;
-
-  return (
-    <Box as="li" sx={{ display: 'grid', gridColumnStart: 2, '::marker': 'none' }}>
-      <Box>
-        <Link
-          sx={{
-            overflow: 'auto',
-            fontWeight: 'semibold',
-            wordWrap: 'break-word',
-            ':link': {
-              color: 'success.fg',
-            },
-            ':visited': {
-              color: 'success.fg',
-            },
-          }}
-          href={link}
-          rel={isTrustedDomain(link) ? undefined : 'nofollow'}>
-          <Text sx={{ wordBreak: 'break-word', marginRight: 1 }}>
-            {title} {domain}
-          </Text>
-          {isAdToExternalLink && <LinkExternalIcon verticalAlign="middle" />}
-        </Link>
-      </Box>
-
-      <Text sx={{ whiteSpace: 'nowrap', fontSize: 0, color: 'neutral.emphasis' }}>
-        Contribuindo com{' '}
-        <Tooltip
-          aria-label={`Autor: ${ad.owner_username}`}
-          direction="nw"
-          sx={{ position: 'absolute', display: 'grid' }}>
-          <Link
-            sx={{ overflow: 'hidden', textOverflow: 'ellipsis', color: 'neutral.emphasis', mr: 2 }}
-            href={`/${ad.owner_username}`}>
-            {ad.owner_username}
-          </Link>
-        </Tooltip>
-      </Text>
-    </Box>
-  );
 }
